requirements.txt
python-pptx==0.6.18
click>=7.1.2
marshmallow==3.6.1